+++ SteakBot.Core/Objects/Enums/MemeResultType.cs
@@ -1,0 +1,8 @@
+﻿namespace SteakBot.Core.Objects.Enums
+{
+	internal enum MemeResultType
+	{
+		Text = 1,
+		Image = 2
+	}
+}
